Moved Lido implementation to interop
