+++ crates/common/src/interop/lido/types.rs
@@ -1,0 +1,15 @@
+use alloy::sol;
+
+sol! {
+    #[allow(missing_docs)]
+    #[sol(rpc)]
+    LidoRegistry,
+    "src/abi/LidoNORegistry.json"
+}
+
+sol! {
+    #[allow(missing_docs)]
+    #[sol(rpc)]
+    LidoCSMRegistry,
+    "src/abi/LidoCSModuleNORegistry.json"
+}
