+++ crates/common/src/interop/mod.rs
@@ -1,1 +1,2 @@
+pub mod lido;
 pub mod ssv;
